Full/django/app/requirements.txt
django
djangorestframework
drf-yasg>=1.21.4
psycopg2>=2.9.0